+++ examples/next/pages/ui/components/ai/ai-conversation/amplify_outputs.js
@@ -1,2 +1,2 @@
-const amplifyOutputs = {};
+import amplifyOutputs from '@environments/ai/gen2/amplify_outputs';
 export default amplifyOutputs;